requirements.txt
aiohttp
dnspython
asyncio
requests
json